rpc: factor out get_committee logic
