+++ crates/sui-rpc-api/src/service/mod.rs
@@ -1,5 +1,6 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
+mod committee;
 mod health;
 mod info;
